+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [Unreleased]
 ### Added
+- Parser settings (CSV, JSON) can be validated/tested by uploading a file and viewing the parsing result immediately ([Merge Request](https://codebase.helmholtz.cloud/ufz-tsm/tsm-orchestration/-/merge_requests/668))
 ### Fixed
 - Dragging columns in overview tables does not change the sorting behavior ([Merge Request](https://codebase.helmholtz.cloud/ufz-tsm/tsm-orchestration/-/merge_requests/700))
 - Correct validation of required parser fields ([Merge Request](https://codebase.helmholtz.cloud/ufz-tsm/tsm-orchestration/-/merge_requests/704))